workshop-webinar/workshops/July-15-2025/customer_support_rag/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
emoji==2.14.1
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.5
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
olefile==0.47
openai==1.95.1
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.15
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.3
unstructured-client==0.38.1
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
